Moved column and constraint serializers into their own files.
