+++ mathesar/api/serializers/constraints.py
@@ -1,0 +1,13 @@
+from rest_framework import serializers
+
+from mathesar.models import Constraint
+
+
+class ConstraintSerializer(serializers.ModelSerializer):
+    name = serializers.CharField(required=False)
+    type = serializers.CharField()
+    columns = serializers.ListField()
+
+    class Meta:
+        model = Constraint
+        fields = ['id', 'name', 'type', 'columns']
