+++ mathesar/api/serializers/serializers.py
@@ -3,11 +3,10 @@ import requests
 from django.urls import reverse
 from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
-from rest_framework.fields import empty
-from rest_framework.settings import api_settings
 
 from mathesar.api.filters import FILTER_OPTIONS_BY_TYPE_IDENTIFIER
-from mathesar.models import Table, Schema, DataFile, Database, Constraint
+from mathesar.api.serializers.columns import SimpleColumnSerializer
+from mathesar.models import Table, Schema, DataFile, Database
 
 
 SUPPORTED_URL_CONTENT_TYPES = {'text/csv', 'text/plain'}
@@ -23,18 +22,6 @@ class ModelNameField(serializers.CharField):
         return value.name
 
 
-class InputValueField(serializers.CharField):
-    """
-    Takes in an arbitrary value. Use to emulate our column and record creation and
-    update endpoints, which handle arbitrary data pulled from request.data
-    """
-    def to_internal_value(self, data):
-        return data
-
-    def to_representation(self, value):
-        return value
-
-
 class SchemaSerializer(serializers.HyperlinkedModelSerializer):
     name = serializers.CharField()
     database = ModelNameField(max_length=128)
@@ -42,91 +29,6 @@ class SchemaSerializer(serializers.HyperlinkedModelSerializer):
     class Meta:
         model = Schema
         fields = ['id', 'name', 'database', 'has_dependencies']
-
-
-class TypeOptionSerializer(serializers.Serializer):
-    precision = serializers.IntegerField(required=False)
-    scale = serializers.IntegerField(required=False)
-
-    def run_validation(self, data=empty):
-        # Ensure that there are no unknown type options passed in.
-        if data is not empty:
-            unknown = set(data) - set(self.fields)
-            if unknown:
-                errors = ['Unknown field: {}'.format(field) for field in unknown]
-                raise serializers.ValidationError({
-                    api_settings.NON_FIELD_ERRORS_KEY: errors,
-                })
-
-        return super(TypeOptionSerializer, self).run_validation(data)
-
-
-class SimpleColumnSerializer(serializers.Serializer):
-    name = serializers.CharField()
-    type = serializers.CharField(source='plain_type')
-    type_options = TypeOptionSerializer(required=False, allow_null=True)
-
-
-class ColumnSerializer(SimpleColumnSerializer):
-    name = serializers.CharField(required=False)
-
-    # From scratch fields
-    type = serializers.CharField(source='plain_type', required=False)
-    nullable = serializers.BooleanField(default=True)
-    primary_key = serializers.BooleanField(default=False)
-
-    # From duplication fields
-    source_column = serializers.IntegerField(required=False, write_only=True)
-    copy_source_data = serializers.BooleanField(default=True, write_only=True)
-    copy_source_constraints = serializers.BooleanField(default=True, write_only=True)
-
-    # Read only fields
-    index = serializers.IntegerField(source='column_index', read_only=True)
-    valid_target_types = serializers.ListField(read_only=True)
-    default = InputValueField(
-        source='default_value', read_only=False, default=None, allow_null=True
-    )
-
-    def validate(self, data):
-        if not self.partial:
-            from_scratch_required_fields = ['name', 'type']
-            from_scratch_specific_fields = ['type', 'nullable', 'primary_key']
-            from_dupe_required_fields = ['source_column']
-            from_dupe_specific_fields = ['source_column', 'copy_source_data',
-                                         'copy_source_constraints']
-
-            # Note that we run validation on self.initial_data, as `data` has defaults
-            # filled in for fields that weren't specified by the request
-            from_scratch_required_all = all([
-                f in self.initial_data for f in from_scratch_required_fields
-            ])
-            from_scratch_specific_in = [
-                f for f in from_scratch_specific_fields if f in self.initial_data
-            ]
-            from_dupe_required_all = all([
-                f in self.initial_data for f in from_dupe_required_fields
-            ])
-            from_dupe_specific_in = [
-                f for f in from_dupe_specific_fields if f in self.initial_data
-            ]
-
-            if len(from_dupe_specific_in) and len(from_scratch_specific_in):
-                raise ValidationError(
-                    f'{from_scratch_specific_in} cannot be passed in if '
-                    f'{from_dupe_specific_in} has also been passed in.'
-                )
-            elif not from_dupe_required_all and not from_scratch_required_all:
-                # We default to from scratch required fields if no fields are passed
-                if len(from_dupe_specific_in) and not len(from_scratch_specific_in):
-                    required_fields = from_dupe_required_fields
-                else:
-                    required_fields = from_scratch_required_fields
-                raise ValidationError({
-                    f: ['This field is required.']
-                    for f in required_fields
-                    if f not in self.initial_data
-                })
-        return data
 
 
 class TableSerializer(serializers.ModelSerializer):
@@ -297,13 +199,3 @@ class DataFileSerializer(serializers.ModelSerializer):
         if content_type not in SUPPORTED_URL_CONTENT_TYPES:
             raise ValidationError(f"URL resource '{content_type}' not a valid type.")
         return url
-
-
-class ConstraintSerializer(serializers.ModelSerializer):
-    name = serializers.CharField(required=False)
-    type = serializers.CharField()
-    columns = serializers.ListField()
-
-    class Meta:
-        model = Constraint
-        fields = ['id', 'name', 'type', 'columns']
