+++ mathesar/api/viewsets/schemas.py
@@ -5,7 +5,7 @@ from rest_framework.response import Response
 
 from mathesar.api.filters import SchemaFilter
 from mathesar.api.pagination import DefaultLimitOffsetPagination
-from mathesar.api.serializers import SchemaSerializer
+from mathesar.api.serializers.serializers import SchemaSerializer
 from mathesar.models import Schema
 from mathesar.utils.schemas import create_schema_and_object
